Add some of Puppeteer's code, and a second scene, to the repo.
This is an excerpt from the sample to draw terrain.  For now it's unmodified.  Wanted to try to get something up and running.

Going to look at switching scenes so we can navigate from the (empty) main menu to the game, and then probably look refactoring the PCX -> Image code a bit, and generally figuring out how to split up our C# across multiple files.  The main motivation being organization, but we should also try not to tie ourselves too closely to Godot this early on, and separating out the business logic from the rendering will help with that.

Also need to figure out how to debug at some point, that's going to be a major drag if we can't debug.