Load up the Civ III background art as the background of the main menu.
Had to remove the button I'd added as I couldn't figure out how to get the background to be below the button if I added the button first.  But that's okay; we'll probably be adding things in a similar fashion anyway.

I'm tempted to try to add an interactable button with PCX graphics to the main menu, since "how do we click on things?" seems like an important thing to understand.