Sorta-kinda set the font color for hover.  Doesn't entirely work though; not sure why.
Also finish clean-up of the file.

We probably want to find a way to group the icon and the text.  We could add an icon to the button directly, but when I tried that, the gap between them was far too narrow, and that also would require figuring out a way to change the icon on hover (maybe you can, needs more research).  But whereas in Civ, hovering on any applies the hover effect on all, here it only affects the button or the label, but not both.

Once more, not really a priority at all, but we'll have a lot of these polish issues along the way so it seems worth diving in a little bit.