Refactor variable names in PCXToGodot.cs (and a little bit ReadPcx.cs) to make it easier to jump into.
I was looking at adding alpha to the menu buttons, and realized I needed to jump to a lower level.  Took me awhile to
realize that the Image variable was not in fact the Image, but the Indices.  So, I renamed that.  Also separated out the
color components, in part because the line was long, but as much to make it easier to see what was being done for alpha
at a glance.  Finally, made a variable out of 255 so it's clearer why we're comparing to it, since 255 can have multiple
meanings when dealing with images.  I'm pretty sure I got the right one, but Puppeteer can change it if not.

Overall, this change may be best motivated by different Firaxis PCXs using different methods of transparency; some use
the LAST_INDEX comparison, but others use separate PCX files for transparency.  It's sort of strange, maybe a relic of
evolving standards as Civ was developed.