Finish setting up the next turn button, including the timer.  It now properly turns on 5 seconds after the player turn starts, and blinks every 0.6 seconds until you end the turn, at which point it turns off.
The 0.6 seconds is a ballpark guess at how often it should be.  I tried recording it and looking at the recording, but didn't find an exact figure; someone who's more video-proficient could likely look at the frames and get a proper figure.  0.6 feels about right though; 1.0 is definitely too slow and 0.5 seemed a bit fast.

Once we have more logic, we won't be just waiting 5 seconds, but looking at units moved and so forth.  Still, it's a start.