Split the lower right info box into its own node.
Adjusted some of its layout params.  It'll still chill out on the lower right side regardless of resolution.

The only thing that's busted is the end turn from the lower right box isn't reaching the game anymore.  I tried to wire it up via signals, but for some reason it isn't working.  Maybe it's because I'm trying to emit a signal from Game.cs in LowerRightInfoBox.cs?  Or maybe it's a silly thing that I'm missing.