Connect up the Advisors button (which I'm not sure I ever used in Civ3, maybe once in 2003) to the Domestic Advisor.  Also move that code into its own method.
Technically it isn't 100% correct forever as that button opens whichever advisor you viewed most recently.  But we only have one advisor so far, so it is effectively correct for now.

As I've been working on making the advisor work well again, it's also been occurring to me that if we have all of the buttons be handled in Game.cs, there is going to be kind of a lot going on there.  Which is less good if we have several people working on it that in my editor where it's just one person and there are never any conflicts... and even there it eventually becomes a bit much.  I think this is part of why WildWeazel was thinking we should have an event based system.  The keyboard listener hears F1 and emits, "show the domestic advisor".  The advisor button is clicked and emits "show the domestic advisor".  Somewhere out in the program, there's an advisor component that's listening for "show the domestic advisor", and when it does, it makes its UI components visible.  That would de-centralize Game.cs as the communication hub for the program.

We do have some decent intentions in this area, but since we're stuck GetParent().GetParent().GetParent()'ing our signals, it isn't working in practice so far.