Made a little progress on the GetParent().GetParent().GetParent() problem.  Followed the guide at https://docs.godotengine.org/en/stable/getting_started/step_by_step/signals.html to add an event from the Advisor node to the C7Game node.  Now, the Domestic Advisor, which is a child of Advisor, only needs to GetParent() once to get to the advisor, and the event from Advisor is wired up to C7Game in C7Game.tscn.
It doesn't really solve the centralization issue yet, but it proves we don't have to bubble up forever to emit signals.  And now that I'm writing this, it occurs to me that we may be able to de-centralize somewhat by wiring up different targets.  I'm not sure the advisor is the best example, as the main UI window may need to be the one that tells that overlay to hide, but the "to" field can be configured in the signals.  The advisor could send a signal to the GameStatus area if that made sense.