Move the texture button declarations into the .tscn instead of the .cs.
Things still work.  This should let me create a script for the advisor button itself, in theory.

One thing I don't like is that we grab the node based on its path in the tree.  It seems probably that we'll re-organize at least some things over time, thus breaking the references.  We should see if there's a more permanent, ID-based way to reference them.