More progress towards decentralized signals.  Now, the advisor button directly emits the show event in the C7Game.tscn file, which is handled by ShowLatestAdvisor in Advisors.cs.  DomesticAdvisor emits the "hide" signal on its parent, Advisors, which is handled in Advisors.cs, as well, and hides the advisor area.
Still need to hook up F1 from Game.cs properly.  I kind of feel like I want a custom signal so we can send in a parameter for which advisor to show.  But since the non-keyboard-related stuff is working, I wanted to commit that now.

One potential concern down the road is that our .tscn is going to wind up having a zillion signals in it.  I guess that's okay for now while we're learning stuff, though, and it kind of is nice that it shows up in the Godot editor visually.