Hook F1 up to the domestic advisor again, with a new custom signal.
I kind of feel like the signal should be from Game to Advisor though, since the keyboard handling is done in Game.  Which I'm still not sure if we should really be handling all our key binds there.  It sort of makes sense, but it would be nice to have a separate keyboard-focused class.