The buttons now load their own textures.
It might be a bit much to have each button have their own script, and load their own things.  I'd originally thought I'd create a control that consists of the three of them and have that do it, and that might still be the better option.  But I wanted to get a feel for this approach, and it does a good job of illustrating how to add new components without having to add any code to Game.cs.

It's also definitely being inefficient with each button re-loading the PCX file from disk.  A control that initializes all of them could address that.  Although I'm just as inclined to add some sort of caching to the PCX loader, when we start caring about PCX loading performance, which I don't think should be today.