Add a hover effect to the AdvisorButton.
This also involved re-writing the getImageFromPCXWithAlphaBlend method to account for the case where the alpha image is smaller than the overall image.  Now, you can pass an offset (default zero) for how many rows offset the alpha will be from the image you are requesting in the imagePCX.

The re-write also eliminates the cropping step through better use of math.

It occurs to me that getting the color at a pixel could also be simplified via a helper method, but we'll leave that for later.