Add several starter unit action buttons.
Alas, I cannot figure out how to get them to be as close to each other as in Civ3.  The Godot editor keeps resetting the margins, which it says it will do based on anchor settings, and I don't know enough about those to overrule it.

I've lined them up in HBoxes inside of a VBox, with the idea being that there will be three HBoxes, one for each row of controls, inside the VBox.  The HBox can then have as many buttons as is appropriate.  This seems like a logical way to do things given the variable number of buttons, but it may not be the best way.  Notably, there is a GridContainer that may work well if we establish a maximum number of buttons in a row.  The reason I'm favoring HBox for now is it seems like it may be easier to add and remove buttons based on whether they're appropriate for the selected unit, and I was hoping that if we just set them to hidden when needed, things would resolve themselves without having to add and remove buttons all the time.

So in short, the UI could use some work, but there are buttons here now, notably the Go To button, which we can start working with to interact with units.