Send an event when a new unit is selected.  I've determined that we cannot send a whole MapUnit across the wire via signals; if we try, we get null on the other end.  string, great.  int, fantastic.  Vector2, sure, why not.  But it doesn't like MapUnit.  I suspect that's cause it's a not very flat object.  Thinking in Java, I'd say it probably has to be something serializable, and we've made no effort to make MapUnit serializable.
Also interesting is that if we create a signal with a MapUnit (or another type it doesn't like), it doesn't even show up in the Node pane in Godot.

So for now I'm just sending the name of the unit.  Which is meh, but I don't want to get hung up on this too long right now.

Also have the back end set all the dummy units to having one movement point so one of 'em will be selected.