Fix up the LowerRightInfoBox's blinky light event, so it works pretty much properly.  Still one GetParent, but that should be okay; it should always live there (and we could re-promote it to a node if we prefer not to have a GetParent).
At this point I'm liking the way the events are looking with regards to end turn.  The TurnEnded event is fired by the main Game.cs, and is listened to by UI items that need to do something when the game ends, such as the GameStatus, which needs to stop blinking.  GameStatus has a separate BlinkyEndTurnButtonPressed signal which is listened to by OnPlayerEndTurn in Game.cs - i.e., the blinky button does not directly end the turn, and does not directly stop blinking.

Of our other two End Turn buttons, one of them has its pressed event connected to _onEndTurnButtonPressed, which will call the end turn method which emits the TurnEnded event if it is the player's turn (which is always, right?).  I'm not sure exactly how the other End Turn button  (the one in the toolbar) works, but it seems to be behaving equivalently.

We should probably reduce the number of End Turn buttons we have.