Wire up the first button to a signal, which is received by Game.cs for now.
For some reason it took a long time to convince Godot to see the signal.  I eventually swapped UnitButtons from a VBoxContainer to a Node, and then it showed up; it stuck around after setting the type back to VBoxContainer.  But at first, no matter what I gave as the recipient function, the debugger said it couldn't find the function.  After reverting everything, and restarting Godot (version 3.4), the signal was visible in the Node area, and I could wire it up, and it worked.

Maybe I just had some local file corrupted, perhaps due to swapping to 3.4, perhaps due to editing it in VSCode, and the revert fixed that.  But at any rate, the signal is being handled correctly now.  It'll need to be send to the engine, but I'll do that later.

(Also probably don't want Game.cs to be the hub for this forever, but I don't know where should be the hub.  It's like we need a Node that isn't really a Node, but is just the thing that talks with the Engine.