Add a ParameterWrapper as suggested by a Godot bug report responder as a work-around to being able to send objects via signals.
I think either this or (if it's convenient in C#) JSON'ifying is the way to go for passing objects.  We could create duplicate types that implement Godot, but that sounds like a decent amount of extra work for no gain, or we could make the data types inherit from Godot.Object, but they really shouldn't logically, and that might cause unexpected issues too.

The lower-right info box now updates the new unit info accordingly.