The spacebar now triggers the "hold" action when a unit is selected.
This got me thinking about how we want to handle keyboard shortcuts for unit actions.  The comment describes my thoughts - we probably want a map of keys to actions, and to go there if no other handlers are called (e.g. an advisor handler).  But I want to add another one of these before deciding on that.