The advisor will now stop a spacebar or enter event if it is open.
We probably want to change this so there's a base class that we extend (implement? does C# have multiple inheritance?),
which handles this, so we don't wind up duplicating it all over.  We also probably want to have one list of keys somewhere.
Or maybe it's all keys by default, overridable to do things relevant to the particular node if we don't want to ignore everything?

Either way it's promising that this is working so far.