Use functional programming to wire up all the UnitButtons instead of just one.
Had been trying to use GetNode and GetParent to get to emitting the signal from UnitButton and it just wasn't having it.  Eventually decided that passing a function would be easier, and perhaps more stable too, and it works pretty well.

Not sure if this is the first choice way to do this in C#, but it seems to work pretty well.