The Domestic Advisor will now appear on a full-screen overlay that blocks input to elements in the background while it is open.
This was surprisingly difficult relative to how small the changeset was.  I tried everything, but it would never center correctly in the CenterContainer (Advisor) that the domestic advisor lived in.  Why wouldn't a center container center?  It beat me.  It turns out the key was changing DomesticAdvisor from a Node2D to a TextureRect, and setting the Texture property on the TextureRect to an image of the size that needed to be centered.  As well as undoing the other things I'd done to center it without the full-screen transparent overlay.

The successful line of thought was derived from studying MainMenu and trying to figure out why its CenterContainer was correctly centering, when Advisor was not.

I somewhat recall Puppeteer changing the LowerRightInfoBox to a TextureRect for what was probably the same or a similar underlying reason.  At the time I was too new to Godot to really follow why.  It still seems sort of weird that CenterContainer only works on TextureRect (or perhaps more accurately, only works on some types of nodes).