Properly put the disband pop-up in the top right.
Spent more time than I'd like trying to convince this to work via anchors and other properties on the new HBox.  Eventually remembered/realized that the problem was it couldn't figure out the size of the DisbandConfirmation and was treating it like it was 2 pixels wide.  Just adding children doesn't cause it to realize what the size of a TextureRect is, it's just like a Node2D at that point.  You have to set its texture to a texture of some size.

So, I set it to a programmatically-created transparent box, of size 10 pixels wider than needed (to create a margin), by the appropriate height.

This "parent node doesn't know the size of the child node" has probably been the most challenging Godot concept for me.