Got the font size changed for the popup.  This really isn't documented anywhere, in particular what the SetFont parameters mean.  Well, it says, "set the theme's font at name in node_point", but what does that mean?  It seems we need to pass "font" as the first parameter and whatever type we want it to apply to as the second.  That's sort of guesswork, but seems to be working.
Also note that if you change the font size of a DynamicFont after you import it, it changes it everywhere.  I set it to size 72 and it affected the End Turn button, even though I haven't applied this theme to that button.  That must be why we have two copies of the font in the Fonts folder, one for each size.