Fix the popup trying to remove the child twice while closing.
That had been caused by the Hide() handler's calling of Hide() at the end, which wound up calling it recursively, which tried removing the now-already-removed child for a second time.

You might ask, maybe the Hide() was already hiding and we could just remove the call to Hide()?  But alas, if we try that, the popup removes the child but stays as an overlay.  Registering our Hide() handler overrode the default, but you still had to call Hide() to hide, and somehow that both called the actual hiding code and was recursive.  Greater Godot minds than I would have to explain how that works.

I also tried just removing the Hide() and setting this.Visible = false in PopupOverlay.cs, but that caused a recursive call too, and thus a double-remove.