Add some load time diagnostics, and start adding some caches for PCX/texture loading.
So far it doesn't help on loading the game scene, which is 1.2-1.3 seconds on my 3.6 GHz i2500k, with Civ3 on a 7200 RPM 2 TB HDD from 2013.  But I also added diagnostics to the DisbandConfirmation, where I'd noticed a lag since Puppeteer added the victorious warrior.  There, it did help.  It went from 225 ms average, to 205 ms with only PCX caching, to 145 ms with texture caching.

We also are not necessarily going through the cached code paths in all areas; I suspect a lot of places are using the lower-level methods still.

Thus I can learn a few things:

 - Caching helps
 - Texture helping helps more than PCX caching.  This implies that the HDD is not the main bottleneck, but rather our PCX-in-memory-to-Godot-texture code path.
 - There's still a lot to be optimized; 145 milliseconds is not very speedy for a popup, and that is testing loading it multiple times in-game (the first time is still about 225 ms).

For now I'm going to focus on the higher level (rather than optimizing the texture creation method, though I may revisit that later), and focus specifically on the DisbandConfirmation to learn more about what's bottlenecking a small area, and see if my learnings there are more generally applicable.