Speed up FLC import by 40%, from a bit over 700 ms to 420 ms, by switching from SetPixel to LoadBmpFromBuffer.
SetPixel was taking up half the time before, about 350 ms out of just over 700.  This new method reduces that time by 80%.

One major call-out about this code is that it uses an unsafe helper function to process the bits, including pointers and pointer arithmetic.  When thinking "how do we create a byte buffer where there are various length primitives included in it?" this seemed like the way to go, as it allows us to treat shorts and ints as shorts and ints by pointer casting.

We might be able to work around it without an unsafe block by tossing the ints/shorts in the header in as multiple bytes, based on their constituent bytes, and using bytes from the palette as-is.  But that introduces a different level of complex reasoning.

Regardless of what we decide about the specifics of this function, it shows that there are real and significant gains to be had by using byte buffers instead of SetPixel, and it was fun to dive back into the BMP format and learn how to write some low-level C-style C#.