Change the default styles for buttons to match Civ's themeing.
Themes aren't too bad once you get the hang of them.  Although they are a bit laggy in the Godot editor, and sometimes I have to restart the editor for Godot to pick up the changes.  Saving them in VSCode seems to cause it to pick up the changes more quickly.