First steps towards drawing a city label.  Basically a proof-of-concept on the blitting part.
Aside from missing a bit of darker shading around the borders of the civ colors, the appearance is fairly faithful; one unknown is exactly what shades of gray and level of transparency are used, but the current choices pass the "in the same ballpark" test.

Remaining notes:
 - The width should be dynamic, and is not currently.
 - The civ color is hard coded
 - The name is hard coded
 - The font is a little bit too big.  It sort of works, but might run into the lower row once that one is added.
 - Haven't added the population.
 - Haven't added capital/spy center indicators.
 - Haven't added what's being produced.

So, there's a lot yet to go.  But it's always good to commit often.