Center the city name and production dynamically.
At this point, the display code is basically done for the time being.  Might tweak the width a tiny bit, will eventually need the spy logo, and want to make the font dynamic so you can use Noto or LSANS.  But it basically works.

Now, I think the focus should be adding more city dynamics, and perhaps civ dynamics (civ colors).  Also curious to set up a preference for whether the civ name is displayed or not (IIRC that's color blind help in game, though it helps anyone who plays a game only to find India, the Iroquois, and someone else all have the same color of purple).