Add mountains and hills to the map.
For now, it always uses the "first"/standalone mountain/hill graphic.  First priority was making sure things lined up like they should on the map.  Which now they do.

As a note, right now we're using the name to identify the terrain type.  While convenient, that is going to break if anyone ever attempts to import a game from a non-English version of Civ, or a mod that renames the terrain.  We could go back to IDs, but I am thinking maybe a "key" that would correspond to the English name would be a good human-readable compromise.  After all, I already switched 7 and 5 for forest/hills while working on this, so it's definitely easy to misremember which type is which integer.