Show different mountain/hill graphics based on a tile's neighbors.
After pondering a few options for how to get a tile's neighbors, I decided to go with letting a tile know who its neighbors are upon load.  This winds up reading pretty nicely, and still feels right inside of an object-oriented context.  Combined with the Tile.NONE object avoiding the need for null checks, it winds up being a nicer version of the latest version of what I have in my editor.

Also switched to grabbing part of a PCX texture, like Flintlock was doing for the base terrain.  I'm not really sure that this is better than the approach I took in my editor, of chopping up the PCX into a bunch of textures (16 in the case of mountains, 81 for base terrains).  One would think that the approach we have here would result in it having to re-extract each mountain from the larger texture a bunch of times, although perhaps Godot has some caching built in for that.  From a readability standpoint, the one thing that tripped me up was that the mountainRectangle takes pixel coordinates for the offset, not multiples of the size Vector2 that it takes as the third parameter.  While this makes sense, it was unexpected after thinking in terms of higher-level mountain offsets for a long time.

It does look like we have a slight gap between hills and mountains that I do not see in the Civ 3 editor; there is also one in my editor, but a smaller, less noticeable one.  Most likely our drawing offsets are ever so slightly off.