Draw random large jungle graphics wherever there is jungle.
This is obviously wrong, but I'm still offline, and don't have any more detailed knowledge of when to use large vs small jungle.  I suspect it's related to neighboring jungles, or perhaps neighboring coast, but more research on that matter is warranted.

This branch is also now branched off of the MountainsAndHillsPartTwo branch, since I appear not to have synced the merge commit from that branch to my computer before going offline, and when I initially created it, it was off an old version of Development.  Hopefully the cable guy will be able to fix the Internet tomorrow and things will sync again.