Display pine forests.  Update import code to set pine forests in the JSON.
I'm kind of wondering if the C7-static-map-save.json path is the way to go.  It's generating huge line diffs, which suggests it's sort of weird for a change set.  It's also kind of a pain to remember to re-run the generator.

It probably is useful because we can bundle it.  Still, I'm thinking it may be less of a pain to bundle traditional Civ resources than to be limited to whatever Godot supports.  We may well want to do that eventually anyway.