Change terrainType to baseTerrainType.
Also change the toString to print both, not just the base.  The previous commit changed the right-click print to use overlay rather than base, as just base wasn't very useful when trying to see if types matched.

Those two changes reinforce my view that we were missing things due to calling the base just "terrainType".  It was too easy to think "terrainType" was what you'd see on the map.