Regenerate the JSON again.  Renaming the variable broke it.
I think after the next commit it'll be time for a refactoring/tech debt phase.  One item is importing actual terrains rather than dummy coding them all.  But I also want to solve the JSON problem in the dev cycle.  It's far from the first time I've been investigating why my code isn't working and finding out it's because I loaded the default map, and the default map needs to be re-generated.  Yet the game can read from a SAV file just fine, and doing so is a sufficient test for whether ImportCiv3 is working as expected.