First part of reading in real terrain types, rather than dummy data.
For now, it just prints out the names and the number of possible resources.  It also shows that it's fetching data in the framework of the BicData, which appears to be a first.

This actually covers the toughest part of the TERR section, where there is a variable data length.  I'm using a terrainNameOffset variable as an offset for the parts of TERR after that variable data length section.

I am pretty sure I'm not doing this the exact way that Puppeteer had envisioned, using RawData instead of the helpers in Civ3 file.  I plan to look at converting it so it can use the convenience methods there next.  But I wanted to get to a working stop point before trying to refactor that in.