Switch to Puppeteer's method of reading in data (except for the complex possible resources area).
It's definitely more readable, especially for strings.  I wonder if it's as performant, using much larger offsets on the whole file, rather than smaller offsets from RawBytes?  Not planning to investigate it now, until we get more high-level metrics to see if it would even be noticeable if we optimize Civ3 import.