Add some basic city production and growth, and have the city labels on the map reflect that status accurately.
So far the production chain is all hard-coded.  But it's still kind of cool to see units get produced, and be able to build another city.  It was also cool seeing that growth and production of a city size consuming unit (Settler) worked flawlessly.

I did notice that forests and marshes are currently free to move across.  Gonna have to look into why that is.