Fix the margin on the Build City popup, post-refactor.
This was supposed to be part of PR #69, but it's been a week since the creator said he'd add it, and it's on a fork so I couldn't add it there, so I'm just doing it here post-merge.

Not sure how that would generally be done on open source projects, but it seems better than letting the PR hang out all year or rejecting the other 98% of it that we want on a technicality.