Unit production in cities now uses unit prototypes, and the new IProducable interface.
Still have some cleanup to do with TurnHandling.cs and CityInteractions.cs; I think there's a little bit too much in CityInteractions.cs right now, and there's almost nothing in TurnHandling.  I also semi-intentionally introduced a parallel paradigm whereby City uses delegates to call back up to CityInteractions for post-build actions... that should arguably be done in TurnHandling instead, which I'd forgotten about while writing that.

I think it's still two steps forward, even with that one step back though.  I like the GetNextItemToBeProduced; with enhancements, it could lead to a foundation for an AI that considers multiple things about what to produce.

As it is, player cities will now have a 1/3 chance of building Galleys instead of Chariots when they are coastal and a Chariot would normally be built.  Unfortunately, galleys will only travel on land for now... but that will be fixed soon enough.