Put city production in a new AI class.
This undoes the last commit's delegate-based model, but I think the separation of concerns is clearer - at least, having the decision of what to build next be in an AI class instead of an engine EntryPoint is an improvement.

It also simplifies City, which had been getting a bit over-complicated for a data class.