Move shield and population cost into IProducable.
I thought about putting population cost in UnitPrototype, which is where it lives in Civ3.  But then I thought, why does it have to be only for units?  What if we want a sacrificial altar to sacrifice a citizen when it's completed?  Or set it negative for a building, and building it would give you a bonus population (maybe ripe for abuse if you sell and re-build the building, but maybe that's a feature, or maybe you make it not sellable like aqueducts).