Start adding some opponents.
The Roadmap doesn't say when we should start adding opponents.  This is probably *slighhhhhtly* too early, since there's no combat/ways to reduce the number of units.  But I'd argue that it should be earlier rather than later for several reasons:

 - Makes the game more interesting.  I've played Civ3 games in "no opponents, just build a civilization all over the whole world!" mode, and it gets boring.  I've never *finished* such a game.
 - Lets us start playing around with AI, which is one of the big uncertainties.  I can only imagine trying to add AI once all the mechanics are in place.
 - Gives a chance of having the AI become halfway decent at the earlier mechanics.  I reckon the chances are better than if at the end we say, "well, we have 4000 features, let's write an AI that tries to use them all".
 - Helps with emergent design of the code.  I've already realized that switching from the flat-file/save-format list of all units, to grouping them by player, probably makes sense.

As of this commit, it's not great yet.  Four players are spawned and they're supposed to all have different colors, but don't.  The player can control everyone's units, but we knew we had a hack for unit control so that's not a surprise; now's the time to address that.