Unit selection now only selects the human player's units.
It's fairly clear that storing units both at a top-level level and at a player level (and at a tile level, etc.) is a nuisance.  I am favoring eliminating the top level (we can always get a flat version from the players if we need it), while keeping copies that are useful, e.g. tiles to see if combat should occur.