Add a settler AI, which the computer is now using to decide on locations for and to found new cities.
It's not super amazing yet, but is functional, using CxxC or CxxxC patterns, with the AI figuring out it can do CxCxC instead of CxxxxC.  Location choice is weighted by terrain.  New settlers only look within about a 4-tile radius of their home city, which is limiting, and the AI doesn't consider settling overseas at this point.  But the AI will start to forge its own empire.

Also enhance GetNextCityName to loop, so it won't crash the game when the AI builds too many cities.