The Settler AI will now consider moving to sites that are close to cities other than its initial city.  There's a penalty for distance, but it's moderate.
This should fix the problem where eventually (200+ turns in) all the spots in its rings were settled and getting scores of zero, so the AI just chose a 0-score spot and settler it anyway, even if it's right next to another city.

As part of this, we now store the list of player-owned cities on the Player object, and pass them in to the evaluation method.

Also tweaked the AI to be a bit more aggressive in sending out settlers.  Probably slightly too much for a real game, but before was too lax, and this is helpful for testing the settler method.