Teach the AI some Really Bad Defense.  The first unit a city builds where there isn't another unit with defensive abilities passing through will become its defender.  This is a slight improvement over just hoping a randomly exploring unit will be in the vicinity.
The AI also plays the fortify animation every turn instead of leaving it fortified, which is obnoxious but somewhat endearing.  It's trying!

Also disband surplus settlers, due to lack of a Build City action.  There's still a problem with the AI not settling everything, but there's such much going on in the logs by the time that happens that I haven't pinpointed why.  We need a logging framework with multiple levels of logging.