The AI now only fortifies defensive units once.
Add some better debugging by right-clicking tiles than what we had before.  This is in part in lieu of the right-click menu in-game, but is also in part a supplement as it now prints not just the tile/terrain, but the units including their GUIDs, and their strategy.

Thanks to this debugging, I've confirmed that the reason the AI is failing to colonize everything is that it winds up with settlers who can't figure out how to navigate around a bay to a selected location on the other side.  They only consider "which of my neighboring tiles is closer as the crow flies?", and when they get to a point where the fastest way for the crow is to fly across water, and every neighboring land tile takes them farther away, they just bounce back and forth between two tiles until the end of time.

I'm not too surprised; I knew we'd run into pathing issues sooner or later, I just didn't realize it would be quite this soon.

Since we know the destination is on the same continent, we should be able to work around this.  Though we can't just say "if all land tiles take me farther away, figure out a route" and then do the current method on the next tile - we'd still bounce back, at least in some cases.  Example: Trying to go from 17, 49 to 23, 49 on the sample map.  The AI wants to go 3 tiles across water, but needs to take a 9-tile route around the bay.  It's only going to get to that midway point if it can store a route.