Fix a couple bugs in the BFS.  Forgot to add the start tile to the distances array with distance zero, and wasn't skipping the start tile in the path, so the unit was moving to the tile it was already on to start with.
Also add a test for part of the BFS algorithm.  For now this is in a traditional C#-style test project, although IMO we should keep the tests in the same project so it's easy to check what we have tests for.  But it is a successful XUnit test, I think our first one using an actual test framework.

Another age-old problem is how best to test non-public methods in C#.  I don't really want to go the "proliferation of classes that call each other" route, but am not sure what a good solution is in C#.