Add a custom menu background.
This is based on Eugène Delacroix's "Knights Duelling", from white I created a Civ background in 2008.  The font work isn't as fancy yet as Adobe Fireworks made it easier to make it fancy than Krita does, but Krita made it easier than GIMP, and I have the Krita file saved so we can modify it and fancify it as we want.  It's 24 MB though so I'll probably upload it somewhere other than Git.