There's an extension for C# integration + debugging in Visual Studio at https://github.com/godotengine/godot-csharp-visualstudio/tree/1.x, but it's a bit bugged, and hasn't worked smoothly since Godot 3.2.3 RC4 (https://github.com/godotengine/godot-csharp-visualstudio/issues/10).  Eventually, I figured out how to get it working by following the guide at https://gist.github.com/31/302bab2630ddf8d90ab4aec0d358b538 .
To install the extension in the first place, you'll have to add the "Mobile Development with .NET" component of Visual Studio, although you can uncheck the optional Android sub-component of that addition.

Note that I'm using Visual Studio 2019 because I'm on Windows 8.1.  However, reports are that even with the work-around to get it working, using the extension breaks IntelliSense on Visual Studio 2022.  Thus, I recommend sticking with Visual Studio 2019 regardless.

Use the "Launch" debug target to start it, and it works.  "Play in Editor" is the default but requires a Godot editor instance to be up, so I'm not sure why you'd prefer it unless you already had it up to use its GUI tools.